Make sure getAbsolutePathToFile returns path to file
Previously, $file->getPublicgetPublic() was used to return the path
for the file (which might return query parameters), however the
absolute path to the file is expected.

The previous behaviour could result in exceptions if an extension
was used which adds query parameters (or anything else) to the public
URL of the file.

Resolves: #58